Trying to fix warnings about type formatting

+++ hanoi.h
@@ -1,7 +1,7 @@
 #ifndef HANOI_H
 #define HANOI_H
 
-extern unsigned long long move_count;
+extern unsigned long move_count;
 
 void hanoi_recursive(int n, char source, char dest, char aux);
 void hanoi_iterative(int n, char source, char aux, char dest);
